+++ core/types/index.ts
@@ -5,3 +5,4 @@ export * from "./Picture";
 export * from "./Project";
 export * from "./ProjectManager";
 export * from "./Status";
+export * from "./VersionConfig";